Server and Clients wait for new Clients to catch up. Null event caught

+++ HelloProcessing/src/common/timelines/Timeline.java
@@ -15,7 +15,7 @@ public class Timeline {
 	Timeline anchor;
 	protected long origin = 0;
 	
-	public Timeline(Timeline anchor, int ticksize){
+	public Timeline(Timeline anchor, double ticksize){
 		this.anchor = anchor;
 		this.ticksize = ticksize;
 		if(anchor!=null)
